resolved requested changes
